+++ src/api.ts
@@ -1,7 +1,8 @@
 import { Api } from "coder/site/src/api/api"
 import fs from "fs/promises"
-import * as https from "https"
 import * as os from "os"
+import { ProxyAgent } from "proxy-agent"
+import { getProxyForUrl } from "proxy-from-env"
 import * as vscode from "vscode"
 import { CertificateError } from "./error"
 import { Storage } from "./storage"
@@ -30,21 +31,27 @@ export async function makeCoderSdk(baseUrl: string, token: string | undefined, s
       config.headers[key] = value
     })
 
-    // Configure TLS.
     const cfg = vscode.workspace.getConfiguration()
     const insecure = Boolean(cfg.get("coder.insecure"))
     const certFile = expandPath(String(cfg.get("coder.tlsCertFile") ?? "").trim())
     const keyFile = expandPath(String(cfg.get("coder.tlsKeyFile") ?? "").trim())
     const caFile = expandPath(String(cfg.get("coder.tlsCaFile") ?? "").trim())
 
-    config.httpsAgent = new https.Agent({
+    // Configure proxy and TLS.
+    const agent = new ProxyAgent({
+      // If the proxy setting exists, we always use it.  Otherwise we follow the
+      // standard environment variables (no_proxy, http_proxy, etc).
+      getProxyForUrl: (url: string) => cfg.get("http.proxy") || getProxyForUrl(url),
       cert: certFile === "" ? undefined : await fs.readFile(certFile),
       key: keyFile === "" ? undefined : await fs.readFile(keyFile),
       ca: caFile === "" ? undefined : await fs.readFile(caFile),
-      // rejectUnauthorized defaults to true, so we need to explicitly set it to false
-      // if we want to allow self-signed certificates.
+      // rejectUnauthorized defaults to true, so we need to explicitly set it to
+      // false if we want to allow self-signed certificates.
       rejectUnauthorized: !insecure,
     })
+
+    config.httpsAgent = agent
+    config.httpAgent = agent
 
     return config
   })