+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Workspace and agent statuses now show in the sidebar. These are updated every
   five seconds.
+- Support http.proxy setting and proxy environment variables.
 
 ## [v1.0.2](https://github.com/coder/vscode-coder/releases/tag/v1.0.2) (2024-06-12)